move Data folder into src
